Code organization and some other stuff
